+++ packages/xo-web/src/common/intl/messages.js
@@ -618,6 +618,7 @@ const messages = {
   startVmOnMissingHostMessage: 'You must select a host',
   recoveryModeLabel: 'Recovery start',
   suspendVmLabel: 'Suspend',
+  pauseVmLabel: 'Pause',
   stopVmLabel: 'Stop',
   forceShutdownVmLabel: 'Force shutdown',
   rebootVmLabel: 'Reboot',
@@ -871,6 +872,7 @@ const messages = {
   powerStateHalted: 'halted',
   powerStateRunning: 'running',
   powerStateSuspended: 'suspended',
+  powerStatePaused: 'paused',
 
   // ----- VM home -----
   vmCurrentStatus: 'Current status:',
@@ -1459,6 +1461,9 @@ const messages = {
   suspendVmsModalTitle: 'Suspend VM{vms, plural, one {} other {s}}',
   suspendVmsModalMessage:
     'Are you sure you want to suspend {vms, number} VM{vms, plural, one {} other {s}}?',
+  pauseVmsModalTitle: 'Pause VM{vms, plural, one {} other {s}}',
+  pauseVmsModalMessage:
+    'Are you sure you want to pause {vms, number} VM{vms, plural, one {} other {s}}?',
   restartVmsModalTitle: 'Restart VM{vms, plural, one {} other {s}}',
   restartVmsModalMessage:
     'Are you sure you want to restart {vms, number} VM{vms, plural, one {} other {s}}?',
@@ -2075,7 +2080,7 @@ const messages = {
   durationFormat:
     '{days, plural, =0 {} one {# day } other {# days }}{hours, plural, =0 {} one {# hour } other {# hours }}{minutes, plural, =0 {} one {# minute } other {# minutes }}{seconds, plural, =0 {} one {# second} other {# seconds}}',
 }
-forEach(messages, function (message, id) {
+forEach(messages, function(message, id) {
   if (isString(message)) {
     messages[id] = {
       id,